feat: MovieApp 클래스 추가, index 로직 분리
